+++ package.json
@@ -6,8 +6,8 @@
     "bundle": "npx projen bundle",
     "bundle:app/download/download.lambda": "npx projen bundle:app/download/download.lambda",
     "bundle:app/download/download.lambda:watch": "npx projen bundle:app/download/download.lambda:watch",
-    "bundle:app/get-form-overview/getFormOverview.lambda": "npx projen bundle:app/get-form-overview/getFormOverview.lambda",
-    "bundle:app/get-form-overview/getFormOverview.lambda:watch": "npx projen bundle:app/get-form-overview/getFormOverview.lambda:watch",
+    "bundle:app/formOverview/get-form-overview/getFormOverview.lambda": "npx projen bundle:app/formOverview/get-form-overview/getFormOverview.lambda",
+    "bundle:app/formOverview/get-form-overview/getFormOverview.lambda:watch": "npx projen bundle:app/formOverview/get-form-overview/getFormOverview.lambda:watch",
     "bundle:app/listSubmissions/listSubmissions.lambda": "npx projen bundle:app/listSubmissions/listSubmissions.lambda",
     "bundle:app/listSubmissions/listSubmissions.lambda:watch": "npx projen bundle:app/listSubmissions/listSubmissions.lambda:watch",
     "bundle:app/submission/submission.lambda": "npx projen bundle:app/submission/submission.lambda",
