+++ src/app/formOverview/get-form-overview/test/formOverviewRequestHandler.test.ts
@@ -1,6 +1,6 @@
 import { ApiGatewayV2Response } from '@gemeentenijmegen/apigateway-http/lib/V2/Response';
-import { DynamoDBDatabase, SubmissionData } from '../../submission/Database';
-import { S3Storage } from '../../submission/Storage';
+import { DynamoDBDatabase, SubmissionData } from '../../../submission/Database';
+import { S3Storage } from '../../../submission/Storage';
 import * as formDefinitionMockSportAanmelden from '../formDefinition/test/mockFormDefinitionAanmeldenSport.json';
 import * as mockFormVolwassen01 from '../formParser/test/subm_raw_volwassene_01.json';
 import { FormOverviewRequestHandler } from '../getFormOverviewRequestHandler';
@@ -10,7 +10,7 @@ let mockS3GetBatch = jest.fn().mockResolvedValue([]);
 let mockS3Store = jest.fn().mockResolvedValue(true);
 let mockDBGetSubmissionsByFormName = jest.fn().mockResolvedValue([]);
 
-jest.mock('../../submission/Storage', () => {
+jest.mock('../../../submission/Storage', () => {
   return {
     S3Storage: jest.fn(() => {
       return {
@@ -21,7 +21,7 @@ jest.mock('../../submission/Storage', () => {
     }),
   };
 });
-jest.mock('../../submission/Database', () => {
+jest.mock('../../../submission/Database', () => {
   return {
     DynamoDBDatabase: jest.fn(() => {
       return {
